fixed SerializeHelper::writeSize back

+++ pvDataApp/misc/serializeHelper.cpp
@@ -33,11 +33,11 @@ namespace epics {
             if(s==(std::size_t)-1) // null    // TODO remove
                 buffer->putByte(-1);
             else if(s<254)
-                buffer->putByte((int8)s);
+                buffer->putByte(static_cast<int8>(s));
             else
             {
                 buffer->putByte(-2);
-                buffer->putInt((int8)s); // (byte)-2 + size
+                buffer->putInt(static_cast<uint32>(s)); // (byte)-2 + size
             }
         }
